toolcall: move message helper under util
