+++ internal/util/message/message_test.go
@@ -7,7 +7,7 @@
 //
 //
 
-package msgutil
+package message
 
 import (
 	"testing"
